Revert "Use size_t in CycleTabs (#1241)"
This reverts commit bcfb6ab8776f9dd1388d333a5f33f2c0e2a1bf31.

+++ src/tsframe.h
@@ -862,8 +862,8 @@ struct TSFrame : wxFrame {
         }
     }
 
-    void CycleTabs(size_t offset = 1) {
-        auto numtabs = notebook->GetPageCount();
+    void CycleTabs(int offset = 1) {
+        auto numtabs = static_cast<int>(notebook->GetPageCount());
         offset = ((offset >= 0) ? 1 : numtabs - 1);  // normalize to non-negative wrt modulo
         notebook->SetSelection((notebook->GetSelection() + offset) % numtabs);
     }
